+++ R/ojo_connect.R
@@ -10,36 +10,46 @@
 #' @param ... Placeholder
 #' @param .global A logical value indicating whether to establish the connection in the global environment or not.
 #' @param .env The environment in which you want the connection stored
+#' @param .pool A logical value indicating whether to use a connection pool from the `{pool}` package, or not. If `.global = TRUE` and a connection object already exists, this argument is ignored regardless of the connection type.
 #'
 #' @export
-#' @returns A database connection object created with `pool::dbPool` and `RPostgres::Postgres()`
+#' @returns A database connection object created with `RPostgres::Postgres()` and either `pool::dbPool` or `DBI::dbConnect` 
 #'
 #' @examples
 #' \dontrun{
 #' ojo_connect()
 #' }
 #' @section Side Effects:
-#' If either the `.global` argument or `rlang::is_interactive` are `TRUE`, an object named `ojo_pool` is created in the package environment.
+#' If either the `.global` argument or `rlang::is_interactive` are `TRUE`, a connection object (named `ojo_con` or `ojo_pool` depending on the `.pool` argument) is created in the package environment.
 #'
 #' @seealso ojo_auth()
 #'
-ojo_connect <- function(..., .admin = FALSE, .global = rlang::is_interactive(), .env = ojo_env()) {
+ojo_connect <- function(..., .admin = FALSE, .global = rlang::is_interactive(), .env = ojo_env(), .pool = FALSE) {
 
   user_type <- if (.admin) "ADMIN" else "DEFAULT"
 
   if (Sys.getenv("OJO_HOST") == "") {
-    rlang::abort("No {tolower(user_type)} configuration for the OJO database was found. Please create one now using `ojo_auth`, or manually, by adding the necessary environment variables with `usethis::edit_r_environ`.")
+    rlang::abort(
+      "No {tolower(user_type)} configuration for the OJO database was found. Please create one now using `ojo_auth`, or manually, by adding the necessary environment variables with `usethis::edit_r_environ`.",
+      use_cli_format = TRUE
+    )
   }
 
-  # Check if pool with correct user already exists and is valid
-  if (.global && exists("ojo_pool", envir = .env)) {
-    db <- get("ojo_pool", envir = .env, inherits = FALSE)
-    if (pool::dbIsValid(db)) {
-      return(db)
+  # Check if object with correct connection type and user already exists and is valid
+  if (.global) {
+    global_conn <- try(
+      get_connection_object(env = .env),
+      silent = TRUE
+    )
+
+    if (!inherits(global_conn, "try-error") && !is.null(global_conn)) {
+      return(global_conn)
     }
   }
 
-  conn <- pool::dbPool(
+  connection_type <- if (.pool) "ojo_pool" else "ojo_con"
+
+  conn_args <- list(
     drv = RPostgres::Postgres(),
     dbname = "ojodb",
     host = Sys.getenv("OJO_HOST"),
@@ -55,13 +65,27 @@ ojo_connect <- function(..., .admin = FALSE, .global = rlang::is_interactive(), 
     ...
   )
 
+  conn_fn <- switch(
+    connection_type,
+    ojo_pool = pool::dbPool,
+    ojo_con = DBI::dbConnect
+  )
+
+  conn <- rlang::exec(conn_fn, !!!conn_args)
+
   if (.global) {
-    assign("ojo_pool", conn, envir = .env)
+    assign(connection_type, conn, envir = .env)
     withr::defer(
       {
-        if(exists("ojo_pool", envir = .env)) {
-          pool::poolClose(.env$ojo_pool)
-          rm("ojo_pool", envir = .env)
+        if (exists(connection_type, envir = .env)) {
+          connection_object <- get(connection_type, envir = .env, inherits = FALSE)
+          if (.pool) {
+            pool::poolClose(connection_object)
+          } else {
+            DBI::dbDisconnect(connection_object)
+          }
+
+          rm(list = connection_type, envir = .env)
         }
       },
       envir = .env